+++ client/components/planning-editor-standalone/profile.ts
@@ -16,6 +16,7 @@ import {
     IAttachmentsFieldConfig,
 } from '../../planning-extension/src/authoring-react-fields/planning-attachments/interfaces';
 import {IProfileSchemaTypeList} from 'interfaces';
+import {getCustomVocabularyFields} from './field-adapters/custom-vocabularies';
 
 function getTextFieldConfig(options: {id: string; label: string, required: boolean}): IAuthoringFieldV2 {
     const editor3ConfigWithoutFormatting: IEditor3Config = {
@@ -60,20 +61,17 @@ function getDateTimeField(options: {id: string; label: string, required: boolean
     return field;
 }
 
-interface IFieldDefinition {
+export interface IFieldDefinition {
     fieldId: string;
     getField: (options: {required: boolean, id: string}) => IAuthoringFieldV2;
     storageAdapter?: {
-        storeValue: (operationalValue) => unknown; // returns stored value
+        storeValue: <T>(item, operationalValue) => T; // returns stored value
         retrieveStoredValue: (storageValue) => unknown; // returns operational value
     };
 }
 
 type IFieldDefinitions = {[fieldId: string]: IFieldDefinition};
 
-export const SUBJECT_PREFIX_ID = 'subject--';
-const getCustomVocabulariesId = (vocabularyId: string) => `${SUBJECT_PREFIX_ID}${vocabularyId}`;
-const getStrippedCustomVocabularyId = (customVocabularyId: string) => customVocabularyId.replace(SUBJECT_PREFIX_ID, '');
 
 export function getFieldDefinitions(): IFieldDefinitions {
     const {gettext} = superdeskApi.localization;
@@ -146,13 +144,16 @@ export function getFieldDefinitions(): IFieldDefinitions {
                 return field;
             },
             storageAdapter: {
-                storeValue: (operationalValue: Array<string>): Array<IVocabularyItem> => {
+                storeValue: (item, operationalValue: Array<string>) => {
                     const vocabulary = superdeskApi.entities.vocabulary.getAll().get('locators');
                     const vocabularyItems = new Map<IVocabularyItem['qcode'], IVocabularyItem>(
                         vocabulary.items.map((item) => [item.qcode, item]),
                     );
 
-                    return operationalValue.map((qcode) => vocabularyItems.get(qcode));
+                    return {
+                        ...item,
+                        place: operationalValue.map((qcode) => vocabularyItems.get(qcode)),
+                    };
                 },
                 retrieveStoredValue: (storageValue: Array<IVocabularyItem>) => storageValue.map(({qcode}) => qcode),
             },
@@ -175,69 +176,10 @@ export function getFieldDefinitions(): IFieldDefinitions {
             },
         }
     ];
-    const planningProfile = planningApi.contentProfiles.get('planning');
-    const customVocabularyIds =
-        (planningProfile.schema?.['custom_vocabularies'] as IProfileSchemaTypeList)?.vocabularies;
 
-    if ((customVocabularyIds?.length ?? 0) > 0) {
-        const allVocabularies = superdeskApi.entities.vocabulary.getAll();
-
-        for (const id of customVocabularyIds) {
-            result.push({
-                fieldId: getCustomVocabulariesId(id),
-                getField: ({required, id: _id}) => {
-                    const fieldConfig: IDropdownConfigVocabulary = {
-                        source: 'vocabulary',
-                        vocabularyId: id,
-                        multiple: true,
-                        required: required,
-                    };
-
-                    const field: IAuthoringFieldV2 = {
-                        id: _id,
-                        name: id,
-                        fieldType: 'dropdown',
-                        fieldConfig: fieldConfig,
-                    };
-
-                    return field;
-                },
-                storageAdapter: {
-                    storeValue: (operationalValue: {
-                        existing: Array<ISubject>;
-                        fieldId: string;
-                        value?: Array<IVocabularyItem['qcode']>;
-                    }) => {
-                        const {existing = [], fieldId, value = []} = operationalValue;
-                        const strippedId = getStrippedCustomVocabularyId(fieldId);
-                        const vocabulary = allVocabularies.get(strippedId);
-                        const vocabItems = vocabulary.items.filter((x) => value?.includes(x.qcode)) ?? [];
-
-                        // Subfield values
-                        const itemsToSubject: Array<ISubject> = vocabItems.map((x) => ({
-                            name: x.name,
-                            qcode: x.qcode,
-                            scheme: vocabulary._id,
-                        }));
-
-                        // Remove values that don't match the "subfield" ID, so there's no item duplication
-                        const restOfValues = existing.filter((x) => x.scheme !== strippedId);
-
-                        return [
-                            ...itemsToSubject,
-                            ...restOfValues,
-                        ];
-                    },
-                    retrieveStoredValue: (storageValue: {fieldId: string; subject: Array<ISubject>;}) => {
-                        const {fieldId, subject = []} = storageValue;
-                        const strippedId = getStrippedCustomVocabularyId(fieldId);
-
-                        return subject.filter((x) => x.scheme === strippedId).map((x) => x.qcode);
-                    },
-                }
-            });
-        }
-    }
+    result.push(
+        ...getCustomVocabularyFields(),
+    );
 
     const resultObj = result.reduce((acc, item) => {
         acc[item.fieldId] = item;
